chore(common): CHECKOUT-6970 Rename package name
