+++ packages/afterpay-integration/project.json
@@ -5,20 +5,14 @@
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
+      "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": [
-          "packages/afterpay-integration/**/*.ts"
-        ]
+        "lintFilePatterns": ["packages/afterpay-integration/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": [
-        "coverage/packages/afterpay-integration"
-      ],
+      "outputs": ["coverage/packages/afterpay-integration"],
       "options": {
         "jestConfig": "packages/afterpay-integration/jest.config.js",
         "passWithNoTests": true
