--- packages/common/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-    displayName: '<package-name>',
-    preset: '../../jest.preset.js',
-    globals: {
-        'ts-jest': {
-            tsconfig: '<rootDir>/tsconfig.spec.json',
-            diagnostics: false,
-        },
-    },
-    setupFilesAfterEnv: ['../../jest-setup.ts'],
-    coverageDirectory: '../../coverage/packages/<package-name>',
-};
