+++ packages/core/project.json
@@ -1,77 +1,77 @@
 {
-    "root": "packages/core",
-    "sourceRoot": "packages/core/src",
-    "projectType": "library",
-    "targets": {
-        "build": {
-            "executor": "@nrwl/workspace:run-commands",
-            "options": {
-                "command": "webpack --mode production"
-            },
-            "dependsOn": [
-                {
-                    "target": "generate",
-                    "projects": "self"
-                },
-                {
-                    "target": "build",
-                    "projects": "dependencies"
-                }
-            ]
+  "root": "packages/core",
+  "sourceRoot": "packages/core/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "command": "webpack --mode production"
+      },
+      "dependsOn": [
+        {
+          "target": "generate",
+          "projects": "self"
         },
-        "dev": {
-            "executor": "@nrwl/workspace:run-commands",
-            "options": {
-                "command": "webpack --mode development --watch"
-            },
-            "dependsOn": [
-                {
-                    "target": "generate",
-                    "projects": "self"
-                },
-                {
-                    "target": "build",
-                    "projects": "dependencies"
-                }
-            ]
-        },
-        "test": {
-            "executor": "@nrwl/jest:jest",
-            "outputs": ["coverage/packages/core"],
-            "options": {
-                "jestConfig": "packages/core/jest.config.js",
-                "passWithNoTests": false
-            },
-            "dependsOn": [
-                {
-                    "target": "generate",
-                    "projects": "self"
-                }
-            ]
-        },
-        "lint": {
-            "executor": "@nrwl/linter:eslint",
-            "outputs": ["{options.outputFile}"],
-            "fix": true,
-            "options": {
-                "lintFilePatterns": ["packages/core/**/*.{ts,tsx}"]
-            },
-            "dependsOn": [
-                {
-                    "target": "generate",
-                    "projects": "self"
-                }
-            ]
-        },
-        "generate": {
-            "executor": "@nrwl/workspace:run-commands",
-            "options": {
-                "commands": [
-                    "npx nx generate @bigcommerce/checkout/workspace-tools:auto-export --projectName=core"
-                ],
-                "parallel": false
-            }
+        {
+          "target": "build",
+          "projects": "dependencies"
         }
+      ]
     },
-    "tags": ["scope:core"]
+    "dev": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "command": "webpack --mode development --watch"
+      },
+      "dependsOn": [
+        {
+          "target": "generate",
+          "projects": "self"
+        },
+        {
+          "target": "build",
+          "projects": "dependencies"
+        }
+      ]
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/packages/core"],
+      "options": {
+        "jestConfig": "packages/core/jest.config.js",
+        "passWithNoTests": false
+      },
+      "dependsOn": [
+        {
+          "target": "generate",
+          "projects": "self"
+        }
+      ]
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "fix": true,
+      "options": {
+        "lintFilePatterns": ["packages/core/**/*.{ts,tsx}"]
+      },
+      "dependsOn": [
+        {
+          "target": "generate",
+          "projects": "self"
+        }
+      ]
+    },
+    "generate": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "commands": [
+          "npx nx generate @bigcommerce/checkout/workspace-tools:auto-export --projectName=core"
+        ],
+        "parallel": false
+      }
+    }
+  },
+  "tags": ["scope:core"]
 }
