+++ packages/google-pay-integration/project.json
@@ -12,9 +12,7 @@
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": [
-        "coverage/packages/google-pay-integration"
-      ],
+      "outputs": ["coverage/packages/google-pay-integration"],
       "options": {
         "jestConfig": "packages/google-pay-integration/jest.config.js"
       }
