+++ packages/hosted-credit-card-integration/project.json
@@ -5,20 +5,14 @@
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
+      "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": [
-          "packages/hosted-credit-card-integration/**/*.ts"
-        ]
+        "lintFilePatterns": ["packages/hosted-credit-card-integration/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": [
-        "coverage/packages/hosted-credit-card-integration"
-      ],
+      "outputs": ["coverage/packages/hosted-credit-card-integration"],
       "options": {
         "jestConfig": "packages/hosted-credit-card-integration/jest.config.js",
         "passWithNoTests": true
