+++ packages/locale/.eslintrc.json
