+++ packages/locale/README.md
@@ -1,0 +1,11 @@
+# locale
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test locale` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint locale` to execute the lint via [ESLint](https://eslint.org/).
