+++ packages/locale/src/TranslatedString/TranslatedString.tsx
@@ -4,12 +4,12 @@ import React, { FunctionComponent } from 'react';
 import { PaymentMethodProps } from '@bigcommerce/checkout/payment-integration-api';
 
 export interface TranslatedStringProps extends Pick<PaymentMethodProps, 'language'> {
-    id: string;
-    data?: TranslationData;
+  id: string;
+  data?: TranslationData;
 }
 
 const TranslatedString: FunctionComponent<TranslatedStringProps> = ({ data, id, language }) => (
-    <>{language.translate(id, data)}</>
+  <>{language.translate(id, data)}</>
 );
 
 export default TranslatedString;
