+++ packages/locale/tsconfig.spec.json
