+++ packages/payment-integration-api/project.json
@@ -12,9 +12,7 @@
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": [
-        "coverage/packages/payment-integration-api"
-      ],
+      "outputs": ["coverage/packages/payment-integration-api"],
       "options": {
         "jestConfig": "packages/payment-integration-api/jest.config.js",
         "passWithNoTests": true
